Split the type propagation plugin into libr/anal/p/tp ##analysis
