+++ libr/anal/meson.build
@@ -44,7 +44,13 @@ r_anal_sources = [
   'p/anal_autoname.c',
   'p/anal_sbpf.c',
   'p/anal_tcc.c',
-  'p/anal_tp.c',
+  'p/tp/trace.c',
+  'p/tp/sizefn.c',
+  'p/tp/types.c',
+  'p/tp/emu.c',
+  'p/tp/match.c',
+  'p/tp/synth.c',
+  'p/tp/plugin.c',
   'p/anal_a2f.c',
   'p/anal_blaze.c',
   'p/anal_gopclntab.c',
